+++ package.json
@@ -17,7 +17,8 @@
     "framer-motion": "^10.12.16",
     "lucide-react": "^0.233.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.14.2"
   },
   "devDependencies": {
     "@types/react": "^18.0.37",
