+++ src/main.tsx
@@ -1,7 +1,11 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import { ChakraProvider, extendTheme } from '@chakra-ui/react'
-import App from './App.tsx'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import { ChakraProvider, extendTheme } from "@chakra-ui/react";
+import App from "./App.tsx";
+import { About } from "./pages/About.tsx";
+import { Landing } from "./pages/Landing.tsx";
+import { Experience } from "./pages/Experience.tsx";
 
 const override = extendTheme({
   fonts: {
@@ -11,19 +15,39 @@ const override = extendTheme({
   },
 
   styles: {
-    global:
-     {
+    global: {
       body: {
-        color: "#221E41"
-      }
-     }
-  }
+        color: "#221E41",
+      },
+    },
+  },
 });
 
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <App />,
+    children : [
+      {
+        path: "/",
+        element: <Landing />,
+      },
+      {
+        path: "/about",
+        element: <About />,
+      },
+      {
+        path: "/experience",
+        element: <Experience />,
+      },
+    ]
+  },
+]);
+
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <ChakraProvider theme={override}>
-      <App />
+      <RouterProvider router={router} />
     </ChakraProvider>
-  </React.StrictMode>,
-)
+  </React.StrictMode>
+);
